2019/8/20 9:23

+++ src/main/java/com/example/demo/dto/QuestionDTO.java
@@ -11,7 +11,7 @@ public class QuestionDTO {
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
+    private long creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
